kleine Änderungen

+++ src/ch/bbbaden/Casino/Menu/Model/ModelBuyCredits.java
@@ -22,7 +22,7 @@ public class ModelBuyCredits {
     private MainApp mainApp;
     
     public void buyCredits(int amount) throws SQLException, ClassNotFoundException, IOException {
-        int currentBalance = Integer.valueOf(viewMenu.getLblCredit().getText());
+        double currentBalance = Double.valueOf(viewMenu.getLblCredit().getText());
         sql.updateCredit(currentBalance + amount, mainApp.getUser().getName());
         mainApp.startMenu();
     }
